break up project into modules
