+++ core/src/lexer/stack.rs
