+++ core/src/stage/codegen/register.rs
