+++ core/src/stage/type_check/scopes.rs
